DistributedLog: Code Cleanup - Removing redundant methods - Fixing discrepancies in exception handling - Most of these were found by IntelliJ Code Inspection
RB_ID=253149

+++ src/main/java/com/twitter/distributedlog/LogSegmentLedgerMetadata.java
@@ -298,7 +298,7 @@ class LogSegmentLedgerMetadata {
         hash = hash * 31 + (int) ledgerId;
         hash = hash * 31 + (int) firstTxId;
         hash = hash * 31 + (int) lastTxId;
-        hash = hash * 31 + (int) version;
+        hash = hash * 31 + version;
         hash = hash * 31 + (int) completionTime;
         return hash;
     }
